requirements.txt
streamlit 
crewai 
python_dotenv 
pydantic 
unstructured 
crewai[google-genai]